+++ apps/web/src/data/portfolio.json
@@ -98,7 +98,7 @@
         "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/resume-1.webp",
         "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/resume-2.webp",
         "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/portfolio-1.webp",
-        "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/portfoilo-2.webp",
+        "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/portfolio-2.webp",
         "https://iwczlmjdxzcitjvcrsvq.supabase.co/storage/v1/object/public/images/portfolio/silverbeen/portfolio-3.webp"
       ]
     },
